Fix microcontroller type for ItsyBitsy M4 Express (#42)
The microcontroller type is wrong: it's not a ATSAMD51J19A but a G19A. The J19A is only in the Metro M4 and Feather M4, according to Adafruit's `board.txt`. The hardware is thus slightly different, having for example a different number of TCs. I'm not entirely sure about the debug section since I'm not using any of those debug devices, but I made what seemed to be the corresponding change. Feel free to correct me if I'm wrong.